Reworked/improved auto-update build/publish tools.
Instead of working on a single 'auto-update' branch,
that needed to be rebased from master branch's HEAD,
we create an orphan branch named after master:HEAD
containing on the update file.

This allows us to keep/delete updates on a per-commit
basis and shows which commit each update is based upon.

Added loads more checking.